move state-based change resolution strategies into views project
