+++ tests/framework/tools.vitruv.framework.views.tests/build.properties
@@ -1,5 +1,6 @@
 bin.includes = .,\
                META-INF/,\
                lib/
-source.. = src
+source.. = src/,\
+           xtend-gen
 output.. = target/classes/
