+++ tests/framework/tools.vitruv.framework.views.tests/src/tools/vitruv/framework/views/changederivation/UmlStateChangeTest.xtend
@@ -1,4 +1,4 @@
-package tools.vitruv.framework.tests.domains
+package tools.vitruv.framework.views.changederivation
 
 import org.junit.jupiter.api.Test
 import static tools.vitruv.testutils.metamodels.UmlMockupCreators.uml
