+++ tests/framework/tools.vitruv.framework.views.tests/src/tools/vitruv/framework/views/impl/StaticSnapshotViewTest.java
